update/ move file location: GroupCard & StoreCard
